requirements.txt
python-telegram-bot==20.7
faster-whisper==1.0.3
requests==2.32.3
pypdf==4.3.1
